add delete the metadata from the database as option in the setup CCBUGS: 393728

+++ core/libs/database/item/imagescanner.cpp
@@ -498,7 +498,8 @@ void ImageScanner::scanFile(ScanMode mode)
     }
     else
     {
-        if (d->scanMode == Rescan && d->scanInfo.id != -1)
+        if (d->scanMode == Rescan && d->scanInfo.id != -1 &&
+            !MetadataSettings::instance()->settings().clearMetadataIfRescan)
         {
             CoreDbAccess().db()->removeAttributesFromImage(d->scanInfo.id);
         }
